refactor(fusion): the claim helpers live with the scope that makes them
`input_error` and `set_output_errors` had exactly one consumer left, and
`panic_message` reads the payload the scope already names as `Panic`.
`failure.rs` was the three of them; they belong beside the type whose
job they are.
