+++ crates/burn-fusion/src/stream/execution/mod.rs
@@ -2,7 +2,6 @@ pub(crate) mod validator;
 
 mod base;
 mod explorer;
-mod failure;
 mod ordering;
 mod policy;
 mod processor;
@@ -14,7 +13,6 @@ pub(crate) use trace::{log_execution_table, op_kind};
 pub use base::*;
 pub use ordering::*;
 
-pub(crate) use failure::*;
 pub(crate) use scope::*;
 
 pub(crate) use explorer::*;
